Substitution (#2)
* Consolidate ref_node and node.

* Add parsing support.

* Get rid of whitespace tokens.

This should dramatically simplify the grammar tests.

We can figure out "lossless" printing when we get there. An alternative
idea is to derive the whitespace from the parent elements / examining
the gaps between child byte ranges.

* Dead end: Parse tree.

The issue is we cannot use the same node in multiple parents. In rust,
values can only have one owner.

Going to switch to using an arena instead, and using ids / references.

* Use id-arena.

* Refactor names and split into files.

* Support formatting in vscode.